# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: FeeDLooP API
  description: |
    FeeDLooP is a feedback collection service for software development companies. It provides APIs for managing projects, collecting feedback through embeddable widgets, and managing bug reports, initiatives, and feedback submissions.

    ## Authentication
    Most endpoints require authentication using NextAuth.js session cookies. The widget submission endpoint is public to allow embedding.

    ## Rate Limiting
    API endpoints implement rate limiting to prevent abuse. Refer to security documentation for specific limits.

    ## CORS
    Different CORS policies are applied based on endpoint:
    - Widget endpoints: Permissive CORS for embedding
    - Dashboard/API endpoints: Restrictive CORS for security

    ## Widget Integration
    The embeddable widget uses the `/api/widget/submit` endpoint to submit feedback from any website. See the Framework Integration Guide for implementation details.

  version: 1.0.0
  contact:
    name: FeeDLooP API Support
    url: https://github.com/your-org/feedloop
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:3000
    description: Development server
  - url: https://localhost:3001
    description: Alternative development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/projects/{id}/reports:
    get:
      tags:
        - Reports
      summary: List project reports
      description: |
        Returns paginated list of reports for a project with filtering and sorting options.
        Supports searching, filtering by type/status/priority, and date range filtering.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
        - name: type
          in: query
          schema:
            type: string
            enum: [bug, initiative, feedback]
          description: Filter by report type
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
          description: Filter by report status
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Filter by priority
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Search in title and description
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter reports created after this date
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter reports created before this date
        - name: assigned_to
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by assigned user
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, priority, status, title]
            default: created_at
          description: Field to sort by
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Paginated list of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportListItem'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
                  filters:
                    type: object
                    description: Applied filters
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Reports
      summary: Create new report
      description: |
        Creates a new report in the project. This is for authenticated dashboard users.
        Widget submissions use the /api/widget/submit endpoint.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetail'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    ReportListItem:
      type: object
      required:
        - id
        - type
        - title
        - status
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique report identifier
        type:
          type: string
          enum: [bug, initiative, feedback]
          description: Type of report
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Report title
        status:
          type: string
          enum: [active, archived]
          description: Current status
        priority:
          type: string
          enum: [low, medium, high, critical]
          description: Priority level
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        attachment_count:
          type: integer
          minimum: 0
          description: Number of attachments
        user_name:
          type: string
          nullable: true
          description: Name of reporter (if provided)
        user_email:
          type: string
          format: email
          nullable: true
          description: Email of reporter (if provided)
    ReportDetail:
      allOf:
        - $ref: '#/components/schemas/ReportListItem'
        - type: object
          required:
            - project_id
            - description
          properties:
            project_id:
              type: string
              format: uuid
              description: Associated project ID
            description:
              type: string
              minLength: 1
              maxLength: 5000
              description: Detailed description
            user_info:
              $ref: '#/components/schemas/UserInfo'
            diagnostic_data:
              $ref: '#/components/schemas/DiagnosticData'
            internal_notes:
              type: string
              maxLength: 2000
              nullable: true
              description: Internal team notes
            resolution_notes:
              type: string
              maxLength: 2000
              nullable: true
              description: Resolution details
            assigned_to:
              type: string
              format: uuid
              nullable: true
              description: Assigned team member ID
            created_by:
              type: string
              format: uuid
              nullable: true
              description: User who created the report (null for widget submissions)
            resolved_at:
              type: string
              format: date-time
              nullable: true
              description: Resolution timestamp
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/AttachmentResponse'
    CreateReportRequest:
      type: object
      required:
        - type
        - title
        - description
      properties:
        type:
          type: string
          enum: [bug, initiative, feedback]
          description: Type of report
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Report title
        description:
          type: string
          minLength: 1
          maxLength: 5000
          description: Detailed description
        priority:
          type: string
          enum: [low, medium, high, critical]
          description: Priority level (defaults to medium)
        user_info:
          $ref: '#/components/schemas/UserInfo'
        diagnostic_data:
          $ref: '#/components/schemas/DiagnosticData'
        internal_notes:
          type: string
          maxLength: 2000
          description: Internal team notes
...
    UserInfo:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          nullable: true
          description: Reporter name
        email:
          type: string
          format: email
          maxLength: 255
          nullable: true
          description: Reporter email
        browser:
          type: string
          maxLength: 200
          nullable: true
          description: Browser information
        os:
          type: string
          maxLength: 100
          nullable: true
          description: Operating system
        url:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
          description: URL where feedback was submitted
        screen_resolution:
          type: string
          pattern: '^\d+x\d+$'
          nullable: true
          description: Screen resolution (WIDTHxHEIGHT)
        ip_address:
          type: string
          nullable: true
          description: IP address of reporter
    DiagnosticData:
      type: object
      properties:
        user_agent:
          type: string
          maxLength: 500
          nullable: true
          description: Browser user agent
        timestamp:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of data collection
        page_url:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
          description: Page URL
        referrer:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
          description: Referrer URL
        console_errors:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 1000
          description: Console error messages
        network_requests:
          type: array
          maxItems: 100
          items:
            type: object
            required:
              - url
              - method
              - status
              - timestamp
            properties:
              url:
                type: string
                maxLength: 2000
              method:
                type: string
                maxLength: 10
              status:
                type: integer
                minimum: 100
                maximum: 599
              timestamp:
                type: string
                format: date-time
              duration:
                type: number
                description: Request duration in milliseconds
        local_storage:
          type: object
          additionalProperties:
            type: string
          description: Local storage data
        session_storage:
          type: object
          additionalProperties:
            type: string
          description: Session storage data
        viewport:
          type: object
          properties:
            width:
              type: integer
              minimum: 1
            height:
              type: integer
              minimum: 1
        device:
          type: object
          properties:
            type:
              type: string
              enum: [desktop, mobile, tablet]
            touch:
              type: boolean
            orientation:
              type: string
              enum: [portrait, landscape]
    AttachmentResponse:
      type: object
      required:
        - id
        - filename
        - content_type
        - size
        - url
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique attachment identifier
        filename:
          type: string
          minLength: 1
          maxLength: 255
          description: Sanitized filename
        original_filename:
          type: string
          description: Original filename before sanitization
        content_type:
          type: string
          description: MIME type
        size:
          type: integer
          minimum: 1
          maximum: 5242880
          description: File size in bytes (max 5MB)
        url:
          type: string
          format: uri
          description: Download URL
        created_at:
          type: string
          format: date-time
          description: Upload timestamp
...
    PaginationInfo:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there are more pages
        has_prev:
          type: boolean
          description: Whether there are previous pages
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Specific error message
          description: Detailed validation errors
        code:
          type: string
          description: Error code for programmatic handling
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
security:
  - sessionAuth: []
...